requirements.txt
# requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
gunicorn==20.1.0

azure-cognitiveservices-speech==1.24.0
azure-ai-vision
azure-ai-textanalytics
playsound==1.2.2
